refactor(tui): separate app frame contracts
